requirements.txt
discord.py>=1.1.0
emoji
aiomysql
cachetools
